LUCENE-9302: Grouping to use long to avoid overflows

+++ lucene/grouping/src/java/org/apache/lucene/search/grouping/BlockGroupingCollector.java
@@ -352,11 +352,11 @@ public class BlockGroupingCollector extends SimpleCollector {
         new TopGroups<>(
             groupSort.getSort(),
             withinGroupSort.getSort(),
-            totalHitCount,
-            totalGroupedHitCount,
+            (long) totalHitCount,
+            (long) totalGroupedHitCount,
             groups,
             maxScore),
-        totalGroupCount);
+        (long) totalGroupCount);
   }
 
   @Override
