+++ lucene/grouping/src/java/org/apache/lucene/search/grouping/GroupingSearch.java
@@ -183,7 +183,7 @@ public class GroupingSearch {
 
     if (allGroups) {
       return new TopGroups(
-          secondPassCollector.getTopGroups(groupDocsOffset), matchingGroups.size());
+          secondPassCollector.getTopGroups(groupDocsOffset), (long) matchingGroups.size());
     } else {
       return secondPassCollector.getTopGroups(groupDocsOffset);
     }
